#277 Add endpoint /point/all that allow to download list of all admin actual event points. Exported to .csv format
