+++ controllers/point/index.js
@@ -1,0 +1,8 @@
+const express = require('express');
+const router = express.Router();
+
+const all = require('./all');
+
+router.use('/all', all);
+
+module.exports = router;
